backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery[redis]==5.3.6

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# PDF Processing
pdfplumber==0.11.0
pdf2image==1.17.0
Pillow==10.2.0
pytesseract==0.3.10

# Audio Processing
pydub==0.25.1

# HTTP Client (Telegram)
httpx==0.27.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
